+++ test/runtests.jl
@@ -43,8 +43,10 @@ end
     include("test_mklpardiso.jl")
 end end
 
-if Pardiso.PARDISO_LOADED[]
-    @testset "pardiso" begin include("test_pardiso.jl") end
-end
+@testset "block" begin include("test_block.jl") end
 
-@testset "parilu0" begin include("test_parilu0.jl") end
+#@testset "parilu0" begin include("test_parilu0.jl") end
+
+if Pardiso.PARDISO_LOADED[]
+     @testset "pardiso" begin include("test_pardiso.jl") end
+end